Create BaseResource and AuthenticatedResource wrappers
